+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Fixed
+
+- fixed UIntValueColumn.asPrimaryKeyIdentity()
+
 ## [0.13.2] - 2024-02-04
 
 ## Changed